// ui/src/app/model/pea.model.ts
export interface ResourceInfo {
  filename?: string
  exists?: boolean
  isDirectory?: boolean
  size?: number
  modified?: number
  md5?: string
}